Change tx_count size in Block to u32
Former-commit-id: 4238f8aafec4049ac5a6b478a42955252917c74e

+++ exonum/src/messages/tests.rs
@@ -274,7 +274,7 @@ fn test_block() {
     let (pub_key, secret_key) = gen_keypair();
     let ts = SystemTime::now();
     let txs = [2];
-    let tx_count = txs.len() as u64;
+    let tx_count = txs.len() as u32;
 
     let content = blockchain::Block::new(blockchain::SCHEMA_MAJOR_VERSION,
                                          0,
